pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoevolve"
version = "0.1.0"
description = "A general-purpose evolutionary optimization framework that uses LLMs as intelligent mutation and recombination operators."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["evolutionary", "optimization", "llm", "mutation", "genetic-algorithm", "beam-search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.39.0"]
openai = ["openai>=1.0"]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]
all = ["anthropic>=0.39.0", "openai>=1.0"]

[project.scripts]
autoevolve = "autoevolve.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
